Fix type casting bug in complete_combustion_to_CO2_H2O
Bug: Line 195 was casting size_t to int before calling oxygen_required_per_mol_fuel()
Fix: Removed unnecessary static_cast<int>(i), pass size_t directly

The function oxygen_required_per_mol_fuel expects std::size_t, not int.
Casting size_t to int is dangerous because:
- int can be negative, size_t cannot
- On 64-bit systems, size_t is 64-bit but int is typically 32-bit
- This could cause incorrect function calls or undefined behavior

Changed:
  double nu_O2 = oxygen_required_per_mol_fuel(static_cast<int>(i));
To:
  double nu_O2 = oxygen_required_per_mol_fuel(i);

This ensures type safety and eliminates potential overflow issues.

+++ src/combustion.cpp
@@ -192,7 +192,7 @@ std::vector<double> complete_combustion_to_CO2_H2O(const std::vector<double>& X,
         if (n[i] <= 0.0) continue;
 
         // Only treat species that actually require O2 (CO2/H2O have requirement=0)
-        double nu_O2 = oxygen_required_per_mol_fuel(static_cast<int>(i));
+        double nu_O2 = oxygen_required_per_mol_fuel(i);
         if (nu_O2 <= 0.0) continue;
 
         const Molecular_Structure& sp = molecular_structures[i];
